Initial commit
started to structure how i want the emu core to look. the idea is too keep in very clean C.

instead of giant source files, split functions across multiple header files, then include those headers in that source file. this allows for the compiler to still inline stuff.

frontend will still be in c++, but that will be worked on much much later on