complete rewrite, see description for more info...
(AR == armwrestler)

added a few arm instructions, basically just enough for AR to get into thumb.

ive added enough thumb instructions for it now switch back into arm, where it
is currently in a loop waiting for vblank, then after that loop, it has another
loop where it waits for vblank to end.

some of the more complex instructions such as any store or loads aren't finished,
but i have asserts() for edge cases or unimpl stuff, so ill add as i need to (or never!).

the rewrite was because i wrote the previous code all in 1 go, without testing it
against any roms, and looking back at it, 90% of the code was broken.

the new arm code is pretty clean, but the thumb code...i just stuck it all in 1
file for now because cba to create new files per instruction group, at least for now.